feat: support command `python -m aerich`
