+++ tests/test_fake.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 import os
 import re
-import shlex
-import subprocess
 import sys
 from pathlib import Path
 
@@ -11,16 +9,7 @@ import pytest
 
 from aerich.ddl.sqlite import SqliteDDL
 from aerich.migrate import Migrate
-from tests._utils import chdir, copy_files
-
-
-def run_shell(command: str, capture_output=True, **kw) -> str:
-    r = subprocess.run(shlex.split(command), capture_output=capture_output)
-    if r.returncode != 0 and r.stderr:
-        return r.stderr.decode()
-    if not r.stdout:
-        return ""
-    return r.stdout.decode()
+from tests._utils import chdir, copy_files, run_shell
 
 
 @pytest.fixture
